couple of fixups for custom footer
