+++ pages/_app.tsx
@@ -12,12 +12,4 @@ function App({ Component, pageProps }) {
   );
 }
 
-
-import '../styles/globals.css'
-
-function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
-}
-
-
 export default App;
