--- styles/globals.css
@@ -1,4 +1,0 @@
-/* Hide the disclaimer link in the sidebar */
-.nextra-sidebar-container a[href="/disclaimer"] {
-  display: none;
-}
